solid design partten

+++ app/Http/Controllers/BudgetHead/BudgetHeadController.php
@@ -6,14 +6,14 @@ use App\Http\Controllers\Controller;
 
 use Illuminate\Http\Request;
 use App\Models\BudgetHead;
-use Mosquitto\Exception;
+use Exception;
 
 class BudgetHeadController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\Response
      */
     public function index()
     {
@@ -24,10 +24,12 @@ class BudgetHeadController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return array|\Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
      */
     public function create(Request $request)
     {
+        abort_unless(auth()->user()->can('create_budget'),403,'you dont have required authorization to this resource');
+
         try {
             $budget=new BudgetHead;
             $budget->budget_title=$request->budget_title;
@@ -81,10 +83,12 @@ class BudgetHeadController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Http\Response|\Illuminate\Routing\Redirector
      */
     public function update(Request $request, $id)
     {
+        abort_unless(auth()->user()->can('edit_budget'),403,'you dont have required authorization to this resource');
+
         try {
             $budget=BudgetHead::find($id);
             $budget->budget_title=$request->budget_title;
@@ -104,11 +108,20 @@ class BudgetHeadController extends Controller
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return array|\Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
      */
     public function destroy($id)
     {
-        BudgetHead::destroy($id);
-        return redirect(route('budget.index'))->with('success','Budget delete Successfully');
+        abort_unless(auth()->user()->can('delete_budget'),403,'you dont have required authorization to this resource');
+        try {
+            BudgetHead::destroy($id);
+            return redirect(route('budget.index'))->with('success','Budget delete Successfully');
+        }catch (Exception $e)
+        {
+            return ["message" => $e->getMessage(),
+                "status" => $e->getCode()
+            ];
+        }
+
     }
 }
